fix schedule

+++ LiteLoader/Kernel/ScheduleAPI.cpp
@@ -21,14 +21,14 @@ public:
     };
     unsigned int taskId;
     TaskType type;
-    unsigned long long leftTime, interval;
+    long long leftTime, interval;
     int count;
     std::function<void(void)> task;
     HMODULE handler;
 
     ScheduleTaskData::ScheduleTaskData(TaskType type, std::function<void(void)> task, unsigned long long delay,
                                        unsigned long long interval, int count, HMODULE handler)
-            : type(type), task(task), leftTime(delay), interval(interval), count(count), taskId(nextTaskId++) ,handler(handler) {}
+            : type(type), task(task), leftTime((long long)delay), interval((long long)interval), count(count), taskId(nextTaskId++) ,handler(handler) {}
 
     inline unsigned int getTaskId() {
         return taskId;
